requirements.txt
openai
gradio
torch
sentence-transformers==2.7.0
langchain
langchain-community
langchain-openai
langchain-text-splitters
python-dotenv